Xcode 5.1対応

+++ RFCViewer/Document.m
@@ -97,7 +97,7 @@
 
 - (NSString *)rfcUrlStringWithIndex:(NSUInteger)index
 {
-    NSString *urlString = [[NSString alloc] initWithFormat:@"%@/rfc%04u.txt", self.baseUrlString, index];
+    NSString *urlString = [[NSString alloc] initWithFormat:@"%@/rfc%04lu.txt", self.baseUrlString, (unsigned long)index];
     return urlString;
 }
